refactor: move out class constructor to the wrapper argument
I didn't liked that the constructor is a part of object of properties
and methods and that it is placed under "initialize" property name. So
now the class constructor should be provided before an object of
properties and methods. Also the set validation rules were updated.

Added more comments to the for better understanding about what's going
on and why.

All unit tests were updated to the new interface and some new
were added.