refactor: explicitly defined encapsulated items over properties
I disliked that encapsulated items can be placed after parent class,
because it looks like a multiple inheritance and the meaning that that
objects are going to be just encapsulated is implicit.
So the place after parent class is potentially reserved for multiple
inheritance and the items for encapsulation is now should be explicitly
defined in the object of properties and methods for a new class under
"Encapsulate" property name.

Update documentation.
Update unit tests.